+++ lib/unleash_proxy_client_flutter.dart
@@ -15,7 +15,7 @@ import 'package:unleash_proxy_client_flutter/unleash_context.dart';
 import 'package:unleash_proxy_client_flutter/variant.dart';
 import 'package:unleash_proxy_client_flutter/metrics.dart';
 
-import 'event_id_generator.dart';
+import 'id_generator.dart';
 import 'http_toggle_client.dart';
 import 'last_update_terms.dart';
 
@@ -50,6 +50,12 @@ class UnleashClient extends EventEmitter {
   /// The name of the app where the Unleash Client is used
   final String appName;
 
+  /// The internal name used by Unleash to count unique SDK connections
+  String connectionId = '';
+
+  // The SDK user agent used by Unleash to know what version is generating traffic
+  final String sdkName;
+
   /// The name of the environment where the Unleash Client is used
   final String environment;
 
@@ -69,7 +75,7 @@ class UnleashClient extends EventEmitter {
   final String Function() sessionIdGenerator;
 
   /// Exposed for testability purposes
-  final String Function() eventIdGenerator;
+  final String Function() idGenerator;
 
   /// Exposed for testability purposes
   final DateTime Function() clock;
@@ -136,8 +142,11 @@ class UnleashClient extends EventEmitter {
       this.refreshInterval = 30,
       this.fetcher = get,
       this.poster = post,
+      // overwrite in tests, do not change in client code
       this.sessionIdGenerator = generateSessionId,
-      this.eventIdGenerator = generateEventId,
+      // overwrite in tests, do not change in client code
+      this.idGenerator = generateId,
+      // overwrite in tests, do not change in client code
       this.clock = DateTime.now,
       this.disableMetrics = false,
       this.storageProvider,
@@ -147,6 +156,8 @@ class UnleashClient extends EventEmitter {
       this.headerName = 'Authorization',
       this.customHeaders = const {},
       this.impressionDataAll = false,
+      // bump on each release, overwrite in tests, do not change in client code
+      this.sdkName = 'unleash-flutter@1.9.0',
       this.experimental}) {
     _init();
     metrics = Metrics(
@@ -157,6 +168,8 @@ class UnleashClient extends EventEmitter {
         clientKey: clientKey,
         disableMetrics: disableMetrics,
         clock: clock,
+        connectionId: connectionId,
+        sdkName: sdkName,
         emit: emit);
     final bootstrap = this.bootstrap;
     if (bootstrap != null) {
@@ -171,6 +184,7 @@ class UnleashClient extends EventEmitter {
   }
 
   Future<void> _init() async {
+    connectionId = idGenerator();
     actualStorageProvider =
         storageProvider ?? await SharedPreferencesStorageProvider.init();
 
@@ -211,6 +225,9 @@ class UnleashClient extends EventEmitter {
       final headers = {
         'Accept': 'application/json',
         'Cache': 'no-cache',
+        'x-unleash-appname': appName,
+        'x-unleash-connection-id': connectionId,
+        'x-unleash-sdk': sdkName,
       };
       headers[headerName] = clientKey;
       headers.addAll(customHeaders);
@@ -475,7 +492,7 @@ class UnleashClient extends EventEmitter {
 
       emit(impressionEvent, {
         'eventType': type,
-        'eventId': eventIdGenerator(),
+        'eventId': idGenerator(),
         'context': contextWithAppName,
         'enabled': enabled,
         'featureName': featureName,