+++ test/unleash_proxy_client_flutter_test.dart
@@ -182,6 +182,8 @@ void main() {
         },
         storageProvider: InMemoryStorageProvider(),
         sessionIdGenerator: generateSessionId,
+        idGenerator: () => '1234',
+        sdkName: 'unleash-flutter@1.0.0',
         fetcher: getMock);
 
     await unleash.start();
@@ -193,6 +195,9 @@ void main() {
         {
           'Accept': 'AcceptType',
           'Cache': 'no-cache',
+          'x-unleash-appname': 'flutter-test',
+          'x-unleash-connection-id': '1234',
+          'x-unleash-sdk': 'unleash-flutter@1.0.0',
           'CustomHeader': 'CustomHeaderValue',
           'X-Custom': 'CustomExtension'
         }
@@ -209,6 +214,8 @@ void main() {
         headerName: 'CustomHeader',
         storageProvider: InMemoryStorageProvider(),
         sessionIdGenerator: generateSessionId,
+        sdkName: 'unleash-flutter@1.0.0',
+        idGenerator: () => '1234',
         fetcher: getMock);
 
     await unleash.start();
@@ -220,6 +227,9 @@ void main() {
         {
           'Accept': 'application/json',
           'Cache': 'no-cache',
+          'x-unleash-appname': 'flutter-test',
+          'x-unleash-connection-id': '1234',
+          'x-unleash-sdk': 'unleash-flutter@1.0.0',
           'CustomHeader': 'proxy-123',
         }
       ]
@@ -1076,6 +1086,8 @@ void main() {
           appName: 'flutter-test',
           refreshInterval: 10,
           sessionIdGenerator: generateSessionId,
+          idGenerator: () => '1234',
+          sdkName: 'unleash-flutter@1.0.0',
           storageProvider: InMemoryStorageProvider(),
           fetcher: getMock);
 
@@ -1089,6 +1101,9 @@ void main() {
           {
             'Accept': 'application/json',
             'Cache': 'no-cache',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
             'Authorization': 'proxy-123',
           }
         ],
@@ -1098,6 +1113,9 @@ void main() {
           {
             'Accept': 'application/json',
             'Cache': 'no-cache',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
             'Authorization': 'proxy-123',
             'If-None-Match': 'ETagValue'
           }
@@ -1116,6 +1134,8 @@ void main() {
           appName: 'flutter-test',
           refreshInterval: 10,
           sessionIdGenerator: generateSessionId,
+          idGenerator: () => '1234',
+          sdkName: 'unleash-flutter@1.0.0',
           storageProvider: InMemoryStorageProvider(),
           fetcher: getMock);
 
@@ -1129,6 +1149,9 @@ void main() {
           {
             'Accept': 'application/json',
             'Cache': 'no-cache',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
             'Authorization': 'proxy-123',
           }
         ],
@@ -1138,6 +1161,9 @@ void main() {
           {
             'Accept': 'application/json',
             'Cache': 'no-cache',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
             'Authorization': 'proxy-123',
           }
         ]
@@ -1256,6 +1282,8 @@ void main() {
           refreshInterval: 10,
           metricsInterval: 10,
           sessionIdGenerator: generateSessionId,
+          idGenerator: () => '1234',
+          sdkName: 'unleash-flutter@1.0.0',
           storageProvider: InMemoryStorageProvider(),
           clock: () => DateTime.utc(2000),
           fetcher: getMock,
@@ -1275,7 +1303,10 @@ void main() {
             'content-type': 'application/json',
             'Accept': 'application/json',
             'Cache': 'no-cache',
-            'Authorization': 'proxy-123'
+            'Authorization': 'proxy-123',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
           },
           payload
         ],
@@ -1299,6 +1330,8 @@ void main() {
           metricsInterval: 10,
           clock: () => DateTime.utc(2000),
           sessionIdGenerator: generateSessionId,
+          idGenerator: () => '1234',
+          sdkName: 'unleash-flutter@1.0.0',
           storageProvider: InMemoryStorageProvider(),
           fetcher: getMock,
           poster: postMock);
@@ -1323,7 +1356,10 @@ void main() {
             'content-type': 'application/json',
             'Accept': 'application/json',
             'Cache': 'no-cache',
-            'Authorization': 'proxy-123'
+            'Authorization': 'proxy-123',
+            'x-unleash-appname': 'flutter-test',
+            'x-unleash-connection-id': '1234',
+            'x-unleash-sdk': 'unleash-flutter@1.0.0',
           },
           payload
         ],
@@ -1573,7 +1609,7 @@ void main() {
         clientKey: 'proxy-123',
         appName: 'flutter-test',
         storageProvider: InMemoryStorageProvider(),
-        eventIdGenerator: () => '1234',
+        idGenerator: () => '1234',
         sessionIdGenerator: () => '5678',
         fetcher: getMock);
 
@@ -1624,7 +1660,7 @@ void main() {
         clientKey: 'proxy-123',
         appName: 'flutter-test',
         storageProvider: InMemoryStorageProvider(),
-        eventIdGenerator: () => '1234',
+        idGenerator: () => '1234',
         sessionIdGenerator: () => '5678',
         fetcher: getMock);
 
@@ -1674,7 +1710,7 @@ void main() {
         clientKey: 'proxy-123',
         appName: 'flutter-test',
         storageProvider: InMemoryStorageProvider(),
-        eventIdGenerator: () => '1234',
+        idGenerator: () => '1234',
         sessionIdGenerator: () => '5678',
         impressionDataAll: true,
         fetcher: getMock);
@@ -1726,7 +1762,7 @@ void main() {
         clientKey: 'proxy-123',
         appName: 'flutter-test',
         storageProvider: InMemoryStorageProvider(),
-        eventIdGenerator: () => '1234',
+        idGenerator: () => '1234',
         sessionIdGenerator: () => '5678',
         impressionDataAll: true,
         fetcher: getMock);